Reorganizo los ejemplos en paquetes.
