--- Circulo.scala
@@ -1,5 +1,0 @@
-class Circulo extends Figura {
-    def dibujar: String = {
-        "Soy un circulo () y soy de color " + color
-    }
-}
