--- Coche.scala
@@ -1,3 +1,0 @@
-class Coche extends PuedeHacerRuido {
-    def hacerRuido(): String = "Bip bip"
-}
