--- HelloWorld.scala
@@ -1,5 +1,0 @@
-object HelloWorld {
-  def main (args: Array[String]) {
-    println("Hello World")
-  }
-}
