--- Mamifero.scala
@@ -1,3 +1,0 @@
-abstract class Mamifero {
-    def mover(): String
-}
