--- Perro.scala
@@ -1,4 +1,0 @@
-class Perro extends Mamifero with PuedeHacerRuido {
-    def mover(): String =  "Yo camino y corro"
-    def hacerRuido(): String = "Guau"
-}
