--- Persona.scala
@@ -1,5 +1,0 @@
-import com.github.nscala_time.time.Imports._
-
-class Persona(fn: DateTime) {
-  val fechaNacimiento = fn
-}
