--- PuedeHacerRuido.scala
@@ -1,3 +1,0 @@
-trait PuedeHacerRuido {
-    def hacerRuido(): String
-}
