--- Rectangulo.scala
@@ -1,5 +1,0 @@
-class Rectangulo extends Figura {
-    def dibujar(): String = {
-        "Soy un rectángulo [ ] y soy de color " + color
-    }
-}
