+++ src/main/scala/ejemplos/fechas/Persona.scala
@@ -1,0 +1,5 @@
+package ejemplos.fechas
+
+import com.github.nscala_time.time.Imports._
+
+class Persona(val fechaNacimiento: DateTime)
