+++ src/main/scala/ejemplos/fechas/PersonaMain.scala
@@ -1,6 +1,8 @@
+package ejemplos.fechas
+
 import com.github.nscala_time.time.Imports._
 
 object PersonaMain extends App {
   val luis = new Persona( new DateTime(1980, 4, 25, 0, 0))
-   println(luis.fechaNacimiento.toString("dd-MM-yyyy"))
+  println(luis.fechaNacimiento.toString("dd-MM-yyyy"))
 }
