+++ src/main/scala/ejemplos/hola/HolaMundo.scala
@@ -1,0 +1,5 @@
+package ejemplos.hola
+
+object HolaMundo extends App {
+  println("Hola mundo")
+}
