+++ src/main/scala/ejemplos/interfaces/Coche.scala
@@ -1,0 +1,5 @@
+package ejemplos.interfaces
+
+class Coche extends PuedeHacerRuido {
+  def hacerRuido(): String = "Bip bip"
+}
