+++ src/main/scala/ejemplos/interfaces/Mamifero.scala
@@ -1,0 +1,5 @@
+package ejemplos.interfaces
+
+abstract class Mamifero {
+  def mover(): String
+}
