+++ src/main/scala/ejemplos/interfaces/Perro.scala
@@ -1,0 +1,6 @@
+package ejemplos.interfaces
+
+class Perro extends Mamifero with PuedeHacerRuido {
+  def mover(): String =  "Yo camino y corro"
+  def hacerRuido(): String = "Guau"
+}
