+++ src/main/scala/ejemplos/interfaces/PuedeHacerRuido.scala
@@ -1,0 +1,5 @@
+package ejemplos.interfaces
+
+trait PuedeHacerRuido {
+  def hacerRuido(): String
+}
