+++ src/main/scala/ejemplos/polimorfismo/Circulo.scala
@@ -1,0 +1,5 @@
+package ejemplos.polimorfismo
+
+class Circulo extends Figura {
+  def dibujar: String = "Soy un circulo () y soy de color " + color
+}
