+++ src/main/scala/ejemplos/polimorfismo/Figura.scala
@@ -1,9 +1,9 @@
+package ejemplos.polimorfismo
+
 abstract class Figura {
     protected var color: String = "negro"
 
-    def cambiarColor(color: String) = {
-        this.color = color
-    }
+    def cambiarColor(color: String) = this.color = color
 
     def dibujar(): String
 }
