+++ src/main/scala/ejemplos/polimorfismo/FigurasTest.scala
@@ -1,0 +1,11 @@
+package ejemplos.polimorfismo
+
+object FigurasTest extends App {
+  val circulo = new Circulo()
+  val rectangulo = new Rectangulo()
+  val lienzo = new Lienzo()
+
+  rectangulo.cambiarColor("rojo")
+
+  println(lienzo.mostrarFiguras(Array(circulo, rectangulo)))
+}
