+++ src/main/scala/ejemplos/polimorfismo/Rectangulo.scala
@@ -1,0 +1,5 @@
+package ejemplos.polimorfismo
+
+class Rectangulo extends Figura {
+  def dibujar: String = "Soy un rectángulo [ ] y soy de color " + color
+}
